Adding Functionality
Hi Kirk, 
I turned the variable "commands" into a LIST. 
I built a function called "showCommands" with two input variables called, "profile" and "cmds"

When the function runs it needs the "net_connect" as the "profile" input and the LIST "commands" as the input for the "cmds" variable. 
Inside the function "showCommands" there is a FOR loop that will rotate through each item in the LIST "commands" and it will PRINT that to screen. 

After the FOR loop is finished, it will issue the disconnect function and the program will terminate.